requirements.txt
sklearn
Pillow
numpy
matplotlib
